Refactor: Move markers and configurators to Util namespace
