+++ Runtime/StateUnit.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using Nopnag.EventBusLib;
 using Nopnag.StateMachineLib.Transition;
+using Nopnag.StateMachineLib.Util;
 using UnityEngine;
 using System.Linq;
