+++ Runtime/Util/TransitionConfigurator.cs.meta
